+++ src/main/java/com/example/demo/repository/ObukaProcedureRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ObukaProcedureRepository extends JpaRepository<ObukaProcedure, int>, JpaSpecificationExecutor<ObukaProcedure> {
+public interface ObukaProcedureRepository extends JpaRepository<ObukaProcedure, Long>, JpaSpecificationExecutor<ObukaProcedure> {
 }
